Narrow mesh dataflow control to u8
Use `u8` for mesh `dataflow` fields and ports across the execute-controller packing path so the interface matches the actual control width. This also removes an unnecessary cast in `ExCtrlMeshCntlPack` and adds a clarifying note on `ExCtrlMeshReq` semantics.

+++ smesh/include/ExCtrlMeshCntlPack.hpp
@@ -55,7 +55,7 @@ class ExCtrlMeshCntlPack : public Component {
   Input(u32, total_rows);
   Input(bit, rs_tag_valid);
   Input(SmeshRsTag, rs_tag);
-  Input(u32, dataflow);
+  Input(u8, dataflow);
   Input(bit, prop);
   Input(u32, shift);
   Input(bit, im2colling);
